import lxd_2.16.orig.tar.gz

+++ lxd/storage.go
@@ -16,7 +16,6 @@ import (
 	"github.com/lxc/lxd/shared/api"
 	"github.com/lxc/lxd/shared/ioprogress"
 	"github.com/lxc/lxd/shared/logger"
-	"github.com/lxc/lxd/shared/version"
 )
 
 // lxdStorageLockMap is a hashmap that allows functions to check whether the
@@ -117,26 +116,29 @@ type storageType int
 
 const (
 	storageTypeBtrfs storageType = iota
-	storageTypeZfs
-	storageTypeLvm
+	storageTypeCeph
 	storageTypeDir
+	storageTypeLvm
 	storageTypeMock
+	storageTypeZfs
 )
 
-var supportedStoragePoolDrivers = []string{"btrfs", "dir", "lvm", "zfs"}
+var supportedStoragePoolDrivers = []string{"btrfs", "ceph", "dir", "lvm", "zfs"}
 
 func storageTypeToString(sType storageType) (string, error) {
 	switch sType {
 	case storageTypeBtrfs:
 		return "btrfs", nil
-	case storageTypeZfs:
-		return "zfs", nil
+	case storageTypeCeph:
+		return "ceph", nil
+	case storageTypeDir:
+		return "dir", nil
 	case storageTypeLvm:
 		return "lvm", nil
 	case storageTypeMock:
 		return "mock", nil
-	case storageTypeDir:
-		return "dir", nil
+	case storageTypeZfs:
+		return "zfs", nil
 	}
 
 	return "", fmt.Errorf("invalid storage type")
@@ -146,14 +148,16 @@ func storageStringToType(sName string) (storageType, error) {
 	switch sName {
 	case "btrfs":
 		return storageTypeBtrfs, nil
-	case "zfs":
-		return storageTypeZfs, nil
+	case "ceph":
+		return storageTypeCeph, nil
+	case "dir":
+		return storageTypeDir, nil
 	case "lvm":
 		return storageTypeLvm, nil
 	case "mock":
 		return storageTypeMock, nil
-	case "dir":
-		return storageTypeDir, nil
+	case "zfs":
+		return storageTypeZfs, nil
 	}
 
 	return -1, fmt.Errorf("invalid storage type name")
@@ -184,7 +188,7 @@ type storage interface {
 	StoragePoolVolumeDelete() error
 	StoragePoolVolumeMount() (bool, error)
 	StoragePoolVolumeUmount() (bool, error)
-	StoragePoolVolumeUpdate(changedConfig []string) error
+	StoragePoolVolumeUpdate(writable *api.StorageVolumePut, changedConfig []string) error
 	GetStoragePoolVolumeWritable() api.StorageVolumePut
 	SetStoragePoolVolumeWritable(writable *api.StorageVolumePut)
 
@@ -193,33 +197,35 @@ type storage interface {
 	ContainerCreate(container container) error
 
 	// ContainerCreateFromImage creates a container from a image.
-	ContainerCreateFromImage(container container, imageFingerprint string) error
-	ContainerCanRestore(container container, sourceContainer container) error
-	ContainerDelete(container container) error
+	ContainerCreateFromImage(c container, fingerprint string) error
+	ContainerCanRestore(target container, source container) error
+	ContainerDelete(c container) error
 	ContainerCopy(target container, source container, containerOnly bool) error
 	ContainerMount(c container) (bool, error)
 	ContainerUmount(name string, path string) (bool, error)
 	ContainerRename(container container, newName string) error
 	ContainerRestore(container container, sourceContainer container) error
-	ContainerSetQuota(container container, size int64) error
 	ContainerGetUsage(container container) (int64, error)
 	GetContainerPoolInfo() (int64, string)
 	ContainerStorageReady(name string) bool
 
-	ContainerSnapshotCreate(snapshotContainer container, sourceContainer container) error
-	ContainerSnapshotDelete(snapshotContainer container) error
-	ContainerSnapshotRename(snapshotContainer container, newName string) error
-	ContainerSnapshotStart(container container) (bool, error)
-	ContainerSnapshotStop(container container) (bool, error)
+	ContainerSnapshotCreate(target container, source container) error
+	ContainerSnapshotDelete(c container) error
+	ContainerSnapshotRename(c container, newName string) error
+	ContainerSnapshotStart(c container) (bool, error)
+	ContainerSnapshotStop(c container) (bool, error)
 
 	// For use in migrating snapshots.
-	ContainerSnapshotCreateEmpty(snapshotContainer container) error
+	ContainerSnapshotCreateEmpty(c container) error
 
 	// Functions dealing with image storage volumes.
 	ImageCreate(fingerprint string) error
 	ImageDelete(fingerprint string) error
 	ImageMount(fingerprint string) (bool, error)
 	ImageUmount(fingerprint string) (bool, error)
+
+	// Storage type agnostic functions.
+	StorageEntitySetQuota(volumeType int, size int64, data interface{}) error
 
 	// Functions dealing with migration.
 	MigrationType() MigrationFSType
@@ -244,8 +250,15 @@ type storage interface {
 	// We leave sending containers which are snapshots of other containers
 	// already present on the target instance as an exercise for the
 	// enterprising developer.
-	MigrationSource(container container, containerOnly bool) (MigrationStorageSourceDriver, error)
-	MigrationSink(live bool, container container, objects []*Snapshot, conn *websocket.Conn, srcIdmap *shared.IdmapSet, op *operation, containerOnly bool) error
+	MigrationSource(c container, containerOnly bool) (MigrationStorageSourceDriver, error)
+	MigrationSink(
+		live bool,
+		c container,
+		objects []*Snapshot,
+		conn *websocket.Conn,
+		srcIdmap *shared.IdmapSet,
+		op *operation,
+		containerOnly bool) error
 }
 
 func storageCoreInit(driver string) (storage, error) {
@@ -269,6 +282,13 @@ func storageCoreInit(driver string) (storage, error) {
 			return nil, err
 		}
 		return &dir, nil
+	case storageTypeCeph:
+		ceph := storageCeph{}
+		err = ceph.StorageCoreInit()
+		if err != nil {
+			return nil, err
+		}
+		return &ceph, nil
 	case storageTypeLvm:
 		lvm := storageLvm{}
 		err = lvm.StorageCoreInit()
@@ -346,6 +366,17 @@ func storageInit(d *Daemon, poolName string, volumeName string, volumeType int) 
 			return nil, err
 		}
 		return &dir, nil
+	case storageTypeCeph:
+		ceph := storageCeph{}
+		ceph.poolID = poolID
+		ceph.pool = pool
+		ceph.volume = volume
+		ceph.d = d
+		err = ceph.StoragePoolInit()
+		if err != nil {
+			return nil, err
+		}
+		return &ceph, nil
 	case storageTypeLvm:
 		lvm := storageLvm{}
 		lvm.poolID = poolID
@@ -432,18 +463,33 @@ func storagePoolVolumeAttachInit(d *Daemon, poolName string, volumeName string, 
 
 	if !reflect.DeepEqual(nextIdmap, lastIdmap) {
 		logger.Debugf("Shifting storage volume")
-		volumeUsedBy, err := storagePoolVolumeUsedByGet(d, volumeName, volumeTypeName)
+		volumeUsedBy, err := storagePoolVolumeUsedByContainersGet(d,
+			volumeName, volumeTypeName)
 		if err != nil {
 			return nil, err
 		}
 
 		if len(volumeUsedBy) > 1 {
-			return nil, fmt.Errorf("idmaps of container and storage volume are not identical")
+			for _, ctName := range volumeUsedBy {
+				ct, err := containerLoadByName(d, ctName)
+				if err != nil {
+					continue
+				}
+
+				ctNextIdmap, err := ct.IdmapSet()
+				if err != nil {
+					return nil, fmt.Errorf("Failed to retrieve idmap of container")
+				}
+
+				if !reflect.DeepEqual(nextIdmap, ctNextIdmap) {
+					return nil, fmt.Errorf("Idmaps of container %v and storage volume %v are not identical", ctNextIdmap, nextIdmap)
+				}
+			}
 		} else if len(volumeUsedBy) == 1 {
 			// If we're the only one who's attached that container
 			// we can shift the storage volume.
 			// I'm not sure if we want some locking here.
-			if volumeUsedBy[0] != fmt.Sprintf("/%s/containers/%s", version.APIVersion, c.Name()) {
+			if volumeUsedBy[0] != c.Name() {
 				return nil, fmt.Errorf("idmaps of container and storage volume are not identical")
 			}
 		}
@@ -594,9 +640,6 @@ func createContainerMountpoint(mountPoint string, mountPointSymlink string, priv
 }
 
 func deleteContainerMountpoint(mountPoint string, mountPointSymlink string, storageTypeName string) error {
-	mntPointSuffix := storageTypeName
-	oldStyleMntPointSymlink := fmt.Sprintf("%s.%s", mountPointSymlink, mntPointSuffix)
-
 	if shared.PathExists(mountPointSymlink) {
 		err := os.Remove(mountPointSymlink)
 		if err != nil {
@@ -604,15 +647,22 @@ func deleteContainerMountpoint(mountPoint string, mountPointSymlink string, stor
 		}
 	}
 
-	if shared.PathExists(oldStyleMntPointSymlink) {
-		err := os.Remove(oldStyleMntPointSymlink)
+	if shared.PathExists(mountPoint) {
+		err := os.Remove(mountPoint)
 		if err != nil {
 			return err
 		}
 	}
 
-	if shared.PathExists(mountPoint) {
-		err := os.Remove(mountPoint)
+	if storageTypeName == "" {
+		return nil
+	}
+
+	mntPointSuffix := storageTypeName
+	oldStyleMntPointSymlink := fmt.Sprintf("%s.%s", mountPointSymlink,
+		mntPointSuffix)
+	if shared.PathExists(oldStyleMntPointSymlink) {
+		err := os.Remove(oldStyleMntPointSymlink)
 		if err != nil {
 			return err
 		}
